feat(stage_l1_combined_syncer): l1Syncer cache, tests
