+++ zk/stages/stage_sequence_execute_test.go
@@ -46,6 +46,8 @@ func TestSpawnSequencingStage(t *testing.T) {
 	err = db.CreateEriDbBuckets(tx)
 	require.NoError(t, err)
 
+	_, db2 := context.Background(), memdb.NewTestDB(t)
+
 	chainID := *uint256.NewInt(1)
 	forkID := uint64(11)
 	latestBatchNumber := uint64(20)
@@ -123,7 +125,7 @@ func TestSpawnSequencingStage(t *testing.T) {
 
 	ethermanMock.EXPECT().BlockByNumber(gomock.Any(), nil).Return(latestL1Block, nil).AnyTimes()
 
-	l1Syncer := syncer.NewL1Syncer(ctx, []syncer.IEtherman{ethermanMock}, l1ContractAddresses, l1ContractTopics, 10, 0, "latest")
+	l1Syncer := syncer.NewL1Syncer(ctx, db2, []syncer.IEtherman{ethermanMock}, l1ContractAddresses, l1ContractTopics, 10, 0, "latest")
 	updater := l1infotree.NewUpdater(&ethconfig.Zk{}, l1Syncer, l1infotree.NewInfoTreeL2RpcSyncer(ctx, &ethconfig.Zk{}))
 
 	cacheMock := cMocks.NewMockCache(mockCtrl)
